Add a client for Virtual Network management
This commit adds entities to generate the (unavailable) XML schema for
virtual network specification, and a client with Get and Set operations
for the currently active subscription.

The pattern for using the client is the same as the powershell version -
get the running config, merge in any desired changes and set the
configuration to the new state. There is no documentation on what the
concurrency constraints are with the underlying Azure REST API, I
suspect it is not safe to call concurrently though can't verify this.