# This is a combination of 3 commits. # The first commit's message is: This enables creating deployments with some additional options. A use case example is that without this code it isn’t possible to create a deployment for a custom virtual network/subnet…
See PR #107 for the complete story on this one.

# This is the 2nd commit message:

Fixing up the tests...

# This is the 3rd commit message:

This enables creating deployments with some additional options. A use
case example is that without this code it isn’t possible to create a
deployment for a custom virtual network/subnet…

See PR #107 for the complete story on this one.