The example file was no longer functional. (#552)
You can see in
[Azure/azure-sdk-for-go#494](https://github.com/Azure/azure-sdk-for-go/issues/494)
that there was a server side error occuring when one executed the
sample. Instead of spending the time fixing this issue, it should be
re-triaged into the other sample improvement items that are being
considered. When ASM samples are generated, they shouldn't share this
repository. Rather, they should belong to the Azure-Samples
organization. [Azure-Samples](https://github.com/Azure-Samples)