Reverting submodule back to contents of storage repo.
glide and other dependency management tools are being broken by the addition of the submodule when you try to target a version before the latest v9.0 which contains the submodule. This is because they first clone the whole repository, then
reset back to the correct version. The submodule creates a conflict that cannot be easily resolved by these tools.