Fix build errors on earlier versions of Go.
Add Go 1.6 and 1.7 to travis.
Refactored some 1.8-specific code into version-specific files.
Updated dependencies.
Disable table batch tests on Go 1.6; this is because the MIME multipart
writer fixed the sort order on headers starting in 1.7 so the recordings
are not comparible between 1.6 and later versions.
Add Go 1.7+ build tags to management package.
Updated README for management package indicating that it's deprecated.