Updating `build` commands in CI to be less chatty.
The way we specified `go build` enumerated each package to be compiled.
This created a lot of noise because folders that didn't contain any Go
source themselves, but had children that did caused an error message.
This way, we'll still get all of the protection of making sure that our
entire repo builds, without the noise.

Also, a couple of formatting changes in profileBuilder.