Updated `tools/generator` to read from literate files (#793)
* Removing generator to start again.

* Beginning of nextGen tools/generator

* Formatting changes.

* Adding gofmt call to generation process

* Adding a description of the value of profiles.

* Adding Build phase and execution time notice

* +Doc for each step of the generation pipeline

* Adding documentation and fixing formatting

* Updating `build` commands in CI to be less chatty.

The way we specified `go build` enumerated each package to be compiled.
This created a lot of noise because folders that didn't contain any Go
source themselves, but had children that did caused an error message.
This way, we'll still get all of the protection of making sure that our
entire repo builds, without the noise.

Also, a couple of formatting changes in profileBuilder.

* Updating AutoRest Log Output location

Also, further documentation improvements

* We can't have nice things.

* Updating managment path in profileBuilder

* Tinkering with profileBuilder regexp

* Adding missing Apache Header.

* Fixing regex error

* Adding Latest Strategy logging

* Fixing SemVer'd package latest comparison

* Adding `go vet` generation step

* Supporting nested RPs

* Overriding generator Usage text.

* Testing and fixing the List strategy

* Updating profileBuilder build tags

Previously, separate files in the profileBuilder had different build
tags. This would cause errors when package `main` didn't have a `main`
function.