Remove hard dependency on GOPATH
The profile builder assumes that GOPATH will be set which might not
always be the case; this invariant has been removed.
Include package name in failure message so it can be investigated.
Moved formatting of package files to happen per package; this way it's
limited to touching files only within the current profile (before it
would touch all files which was confusing).