Updates to apidiff tool
To support porting non-breaking changes the generated report now
includes the collection of packages impacted by commit along with a list
of commits that introduce breaking changes.
Moved the changes to a "deltas" field in the report; each entry is an
object described as "base:target" commit range.
Added support for report generation based on a sequence of commits, see
the help documentation for further details on how this works.