Add picker tool for cherry-picking commits
The picker tool is used to cherry-pick non-breaking changes from one
branch to another. Note that this tool will mutate the state of your
working tree.
Added ExecPackagesCmd() for programmatically invoking the packages
command and exported the type of the report it returns.
Added Cherry() and CherryPick() to the repo package.