Making Examples Executable Tests (#27)
* Moving scheduledMessage example to increase visibility

* Removing reference to scheduledMessage example in README

* Moving helloworld example

* Fixing build errors after rebase

* Moving opentracing example

* Fix opentracing/readme.md execution instructions

* Removing extraneous reference to example that has been moved

* Updating SendReceiveScheduledMessage timeline.

Previously, the message was scheduled to be delivered 15 seconds in the future. Service Bus has a schedule granularity more along the lines of a 1-minute window. For that reason, I update this test to take 3 minutes. From there, I had to update some of the other timeout logic in the test harness to allow a test that took more than 1 minute.

* Removing unused const

* Removing examples section from Makefile now that they are incorporated as tests