Changing Receive and ReceiveOne to be blocking calls.
Fundamentally, I'm making this change because of advise that @bketelsen gave me. To paraphrase him, "never design your API to have non-blocking calls. It's not the Go way." However, to back that up, it's been something that bit me several times while I was working with this Service Bus library.