Introduces a Go wrapper for the Keyvault client.
The Keyvault client in the SDK is a translation from the REST API to a Go library.  Because of this it misses out on following Go style and providing needed methods for a Go programmer for practical use cases.

Due to the nature of how the Keyvault service is written, the lack of keyvault internals documentation, errors that provide no value from the service and common needs for Go programmers the REST service can be frustrating.

This attempts to address those issues by wrapping the REST client inside a wrapper that provides a more Go like interface.  It attempts to address common and frustrating errors such as not having a TenantID by not allowing the New() constructor to complete.  It wraps known errors with messges on common causes.

It provides common operations like retrieving the SSL certificate needed for http.Server that is almost impossible to figure out from the Keyvault documentation.  And PEM decoding for inspecting the certificate.

It defines its own types separate from the REST internals as to prevent the user from having to refer to it.  This also has the benefit of removing most pointers which should easy garbage collection and removes many fields that were not populated by the REST service or not required.

Finally it provides a mock library for writing hermetic tests.  This removes the need for the Keyvault service when writing tests.

This is an initial version containing only read methods to the service and not write methods.