Refactor profileBuilder tool
Simplified the code and in the process uncovered several bugs where an
operation would fail but the error was ignored.
Removed dependency on GOPATH.
For the `list` command changed the input format to JSON; this is
required for upcoming changes to support hybrid profiles.
Removed dependency on viper as it wasn't used and cluttered up the code.
The content in aliasPackage.go is a copy from an external dependency so
the core logic is the same but simplified.
Removed all default values and made their flags required.
Properly support packages that are versioned like `v1`.
The sort order for aliased content has been fixed once and for all.