chore: renormalize CRLF vs LF
088b867aad3db160bb83706fe98d4b2c7b734a8f added the .gitattributes so the repo
wouldn't continue to gain CRLF problems. But the CODEOWNERS does not appear in
the latest branch, nor did it appear in the repo when that PR tried to fix things.

I can only find CODEOWNERS in master, and so this commit is against master. If
CODEOWNERS comes from someother branch I'd be happy to fix this there. But latest
doesn't seem to be the place.