Ensure per-try timeout is properly cancelled
Explicitly call cancel the per-try timeout when the response body has
been read/closed by the body download policy.
When the response body is returned to the caller for reading/closing,
wrap it in a responseBodyReader that will cancel the timeout when the
body is closed.
Logger.Should() will return false if no listener is set.