Fix cancellation of per try timeout
Per try timeout is used to ensure that an HTTP operation doesn't take
too long, e.g. that a GET on some URL doesn't take an inordinant amount
of time.
Once the HTTP request returns, the per try timeout should be cancelled,
not when the response has been read to completion.