Refactor azcore public surface area
The request and transport interfaces have been refactored to align with
the patterns in the standard library.
NewRequest() now uses http.NewRequestWithContext() and performs
additional validation.  It requires a context parameter.
The Policy and Transport interfaces have had their ctx parameter removed
as the context is associated with the underlying http.Request.
Pipeline.Do() will validate the HTTP request before sending it through
the pipeline, avoiding retrying on an malformed request.
The Retrier interface has been replaced with the NonRetriableError
interface, and the retry policy updated to test for this.
Request.SetBody() now requires a content type parameter for setting the
request's MIME type.