azidentity: bearer token policy error out when token expired
If the elected goroutine keeps failing to refresh the token, until it exceeds the time window (now is set to 2min). Then it doesn't make sense to still allow other goroutines to reuse the existing (but now expired) token to pass along the pipeline. Instead, the policy shall return with an `AuthenticationFailedError`.