Fix NullValue for maps and slices
Using reflect.New created a pointer to a zero-value underlying type.
For maps and slices this meant nil which lead to incorrect comparisons.
Instead, create an empty map which ensures a unique address.
For slices, create a slice with a length of one to avoid an apparent
optimization with zero-length slices.