Fixing the issue with 404's not actually failing fast.
The main issue is that the negotiateClaim() function was doing it's own
retries and it wasn't properly handling 404 as a "fatal" error.

Rather than continuing to patch the 'internal' version I'm just moving it
all into the the azservicebus package and leveraging the the normal retry
and error recovery. I stopped short of doing a full change (to make it so
we can hold onto an RPC link for claims negotiation).

Some supporting changes:
- mgmtError is now rpcError and is just generically the "request/response"
  error type. Also, the actual decision of creating an rpcError is built
  in to rpcLink.RPC(), rather than being done by callers.
- Small change to make NewRPCLink take an args structure and remove the
  options that weren't used.
- Removed 'tab' calls in code I'm modifying since we're going to be
  replacing them/moving them anyways.