Handle 410 properly for session and non-session based links.
- For non-sessions links retrying will never work - the lock is lost and can't come back.
- For _session_ links it's possible to get the lock after you've lost it since you can basically "recover" the lock by it's name. This comes into play mostly when you lose the lock because of a server-initiated detach.

Most of this is just cleaning up the code so I don't have to add yet another argument to the parameters and can just have a simple struct with names, which helps with readability. I've also made it so the function that amqpLinks uses to determine if it should retry is replaceable.

To keep the refactor size reasonable I didn't take it this _all_ the way through but it's reasonable as it is now.