Merge arm/runtime/poller.go into runtime/poller.go
LRO refactoring has alleviated the need for separation of ARM and
data-plane pollers so the content has been merged into the runtime
package.  The duplicated content under arm/runtime will be removed in a
future release.
The arm/internal/pollers/loc package was renamed to armloc to avoid
colliding with the existing data-plane package.  These will need to be
consolidated in a separate change.
No functional changes to the relocated, internal, pollers.