Add NewRecordingWithOptions constructor
The NewRecording constructor defaults to a poor location for the
recording files and provides no facilities to override it.
Added a new constructor with options for specifying a custom recordings
directory.  The default is to now place the recordings under the
testdata directory.  The old constructor maintains its poor behavior.