CLI Credential: Avoid calling Azure CLI via cmd.exe/sh
It is not necessary to invoke Azure CLI via either `cmd.exe` or `sh`, as itself is
executable and has proper interpreter settings. This avoids some
unneeded assumptions to the running environment, e.g. in Windows it is
then not needed to always ensure `cmd.exe` is available and correctly
set in the %PATH% (see: https://github.com/Azure/aztfy/issues/119).

Also, the working directory for the spawned process is also not
needed to set. By default, it uses the current working directory, which
is just fine.