[azservicebus] Some naming changes for raw AMQP message support for consistency with other stacks #18430
Some naming changes to be more consistent with other track 2 SDKs as well as the AMQP spec:

Rename the AMQPMessage to be AMQPAnnotatedMessage (track 2 consistency, also matches terminology in the spec)
Create an AMQPAnnotatedMessageBody type to house the .Value, .Sequence and .Data fields. This matches more with the logical format of an AMQP message, where the Body is comprised of sections (Data, Value, Sequence), rather than them just being free-floating fields on a message.