[azservicebus] Fixing bug where cancelling link creation could leak resources (#18479)
Attempting to shoe-horn the draining of the channel into the same goroutine doesn't work - it leads to a potential race condition or can lead to a blocking condition identified by @franklinkim in #18477.

The simplest fix is just to treat the original goroutine as a cancellable function call and then handle the return value logic inside of the main func instead. This ensures we always drain the 'done' channel, even if we decide to cancel early.

NOTE, this also unearthed a bug where we were sometimes passing (amqp.<type>, nil) instead of (<interface>, nil) in the amqp wrappers. This was fixed, along with the spot that was unintentionally using it even when there was an error.