A rather big change.
Changing Receiver.ReceiveMessages() from the existing implementation,
which uses AMQP's flow+drain to doing amqp.Release() on messages that
arrive when there is not an active listener.

It seems particularly hard for our AMQP stack to properly recover if we
drain and then have to cancel (to meet time constraints from the user's
side). This leaves the link in a state where we excessively lock
messages, etc...

Switching over to simply releasing mesages when nobody is actively
receiving allows us to avoid waiting for the ACK (faster exit) as well
as allowing us to keep the link alive for longer, which also avoids
some scenarios where the link would get closed.

There are also quite a few changes to the stress tests to make them
signal on error more reliably as well.