Making the recovery logic for connections simpler by having them only be responsible for recovering themselves, not all other links.
Prior to this we'd do a blanket CloseLinks() for the first recovery that happened to reboot the connection. Since each of these recoveries has to happen for each link anyways it's a bit simpler to let them each recover individually.

This also preps us for the receiver redirect feature, where we can't assume that all the links are actually part of the same connection.

Al;so, added a bunch of tests for this.