It's becoming more clear that cancellation doesn't mesh well with Close() in that it might leave links open with no good way to clean up.
In general, leaks are bad, but in some cases, like the $cbs link, it's particularly bad. The $cbs link can't be opened simultaneously - you get one per connection. So if you cancelled when it was attempting to close the link you'd leak it, and be locked out from that point on when attempting to create a new link.