Adding a timeout so link closures have a maximum of 60 seconds before they time out. This fixes an issue that came up when I made a fix in the last release to not timeout on close (since that can leave us in an inconsistent state). This surfaced an interesting problem where customers do see Close() hang.
So this adds in a timeout instead, which will cause us to close the
connection to avoid inconsistency and also to recover overall since it's
likely there's a problem that will require a more thorough reset than a
simple link detach/reattach would give us.

As part of this I also created some more thorough mocks that test more
of the connection/open behavior as well as links, ensuring that we do
cleanup and end up in the state that we want. It's not 100% complete but
it's good enough for this kind of testing and can be expanded.