Some minor refactoring of azcore
Moved ARM resource ID types and helpers to an internal package so they
can be leveraged elsewhere.  The public versions are type-aliased to the
matching internal implementations.
policyFunc has moved to the shared package so it can be used elsewhere.
Sanitizing of URLs has been split into its own function.
Moved a func-local constant to the shared package.