[azservicebus,azeventhubs] Fixing issue where $cbs link being incompletely closed prevented the connection from recovering (#20334)
You're only allowed to have one Sender+Receiver for the $cbs endpoint. In some cases, where there are connection failures or link cleanup is cancelled it's possible for the link to remain open. The only way to fix this is to restart the connection.

Fixes #19504
Likely related to dapr/components-contrib/issues/2617