[azeventhubs] Fixing some authentication issues (#20450)
Making the same changes for authentication errors as I did in PR#20447, and adding in some new test cases for both packages.
    
Some of the tests required me to update the ARM template. I converted it to a bicep template just to make that simpler to understand, overall, but the only real change is that I generate two more connection strings with limited privileges for testing auth errors.