Okay - I think I got it.
So the main issue here is that I'm relying on the GetEventHubProperties to be immediately accurate when I start my test. This works great most of the time because, AFAICT, if you receive a message the properties will also be accurate.

HOWEVER, in some tests I send a message but don't receive it. So it's possible it's not yet available.

I think what happened is some of the tests (specifically the AMQPLinks tests) were sending to an event hub but never receiving it, so they'd exit quickly. Then one of my other tests would come around and start but catch it in between the message actually becoming available, resulting in weird results.

Long story short: the solution is just to create our own little event hub just for the AMQPLinks tests since I do like they actually _do_ something, and we can just ignore their results.