Refinement of fake package per API review
Added an HTTP status code parameter to to methods that return a
response.  For methods that already had such a paraemter, it was
relocated to be the first parameter.
Removed Responder[T].SetHeader method, replacing the functionality with
an http.Header field in SetResponseOptions.
Cleaned up some wording in the server API doc comments.