[azeventhubs] Pass in a cloud when we use the ARM client. (#20890)
We use an ARM client in the azeventhubs test so (unlike other clients) we need to pass in an explicit Cloud value.

This is part of why we were seeing weekly failures with #19744.

There is another issue where the connectivity seems to be very unreliable between our live tests and the cloud itself. EngSys is investigating that.