Adding in the CloudEvent type (#21027)
The `CloudEvent` type implements the CloudEvent from https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md. This same format (and type) is used in the upcoming EventGrid SDK and is also intended to be usable with other messaging packages as well.

Part of the fix for #20970 (remainder comes in a PR for Event Grid)