[azopenai] Make the deployment ID a per-request parameter, rather than a constructor parameter. (#21223)
A rather big change.

Prior to this we required the deployment ID to be passed as a parameter when constructing the client. This had some advantages in that it matched the underlying physical properties for Azure (it's a route parameter, not a request parameteR), but it broke the de-facto conventions established by OpenAI, where the model is considered a request parameter.

This is now changed, with the one concession being that the `Model` field has been renamed to `DeploymentID`, to match with the other Azure SDKs.