Adding in SystemEvents as a separate package.
I've ported in the tests from the .NET Event Grid package, and made any adjustments that seem sensible to catch up
to some of the naming changes. There are still a few lingering ones that need to get looked at (where they're automatically
doing number conversation in cases where it actually comes back as a string or 'n/a').