Custom code and tests.
The non-autogenerated code is there to work around the way that `union` translates to swagger - it typically just produces the leaf objects but the union itself is empty. model_custom.go are manual representations of the unions, themselves.

I've also addressed some linter errors, removing log.Fatal's from the examples.