azcontainerregistry: decouple token exchange from request specifics
I would like to propose a generally-useful token cache in a future
commit. Managing the refresh tokens and exchanging them for access
tokens does not need to rely specifically on the authentication policy
or the specific client request being made, so this commit decouples the
two concerns, thereby enabling the future refactor to be simpler.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>