Fixing a customer issue where they cancel in the middle of our retry loop and prevent our recovery from happening.
This simplifies the overall logic and should be just fine - some of the cases it was trying to handle were mostly related to idleness, and trying to move the bulk of the failures _before_ the `fn` callback, but this should be more or less equivalent while also handling the error immediately. I've got a couple of tests to ensure that some of the important bits are there (partial failure still leads to link recovery).

Also, updated stress tests to include pipefail, which makes sure that, if the underlying test fails, the script _exits_ instead of potentially reporting okay because `tee` happened to work :) Also, the stress tests were incorrectly using the test credential, and we can't/don't use that in stress testing.