Adding in TokenCredential and APIKey testing paths for assistants.
It looks like our recent subscription migration + tokenkeycredential is failing to authorize us properly. Filed an issue to look at it. The code works just fine with AzureCLICredential and worked up until we migrated to the new subscription so we'll have to dig into this further to find the root cause.