[azeventhubs] Adding support for Geo-Replication in Event Hubs (#24477)
As part of this we have to make a breaking change in the API (Offset, which was previously returned as an integer, is now a string). This field is not quite useful for users, but was exposed as it's used as an opaque starter value for checkpoints. It's always been a string internally so no checkpoint upgrades/changes are required, and we expect most users will be able to just transition to using this without any issues.

The changes are mostly:
- Update the interfaces/types to use strings instead of ints
- Update internals to handle an edge case to fallback to a hardcoded position if we can't find a valid checkpoint

And then doing import updates all over since this is now a /v2.